Add process() to replace new Process
The signature of `Process::__construct()` differs based on the version
of the Symfony Process library, and this will cause server commands to
fail if there is a mismatch between the version local-vip assumes and
the version bundled with Composer.

process() handles switching between the different signature types
(string vs array) which should allow for correct behavior in all
environments so long as `$this->process()` is used instead of `new
Process()`.